Add files via upload
Based on the csv generated by ddrScrape.py, this script goes into each shopping center's web page, finds the Property Manager div, and pulls the name and phone number of the property manager out. It then returns a csv with all of the property info, including the property manager's name and phone number.